Plugs Inspector app made much more advanced.
Displays .Net Classes which have been plugged in a list. On selecting one, it displays the Cosmos classes that plug that .Net class and all the plugged/unplugged methods (including private/hidden methods).
You can select a plugged method and run it with all Null parameters to see what it does.